media-ingestion/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
boto3==1.28.44
kafka-python==2.0.2
python-multipart==0.0.6
